parsing base + sphere ok | leaks ok | leaks on gnl

+++ includes/sub_includes/objects.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   objects.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpoder <rpoder@student.42.fr>              +#+  +:+       +#+        */
+/*   By: margot <margot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/30 14:09:51 by rpoder            #+#    #+#             */
-/*   Updated: 2022/11/20 20:10:14 by rpoder           ###   ########.fr       */
+/*   Updated: 2022/11/22 22:23:19 by margot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@ t_intersections		get_object_intersections(t_object *obj, t_ray ray);
 t_tuple				normal_at(t_object *object, t_tuple world_point);
 
 /* sphere.c------------------------------------------------------------------ */
-static t_matrix4	compute_parsing_sphere_transform_m(t_tuple origin, int radius);
-t_object			*create_sphere(t_data *data, t_tuple origin, int radius);
+static t_matrix4	compute_parsing_sphere_transform_m(t_tuple origin, double radius);
+t_object			*create_sphere(t_data *data, t_tuple origin, double radius);
 t_intersections		get_sphere_intersections(t_object *sphere, t_ray ray);
 t_tuple				sphere_normal_at(t_object *sphere, t_tuple object_point);
 
@@ -32,14 +32,14 @@ t_tuple				cube_normal_at(t_object *cube, t_tuple point);
 t_intersections		get_cube_intersections(t_object *cube, t_ray ray);
 
 /* cylinder.c---------------------------------------------------------------- */
-t_object			*create_cylinder(t_data *data, t_tuple origin, int radius);
+t_object			*create_cylinder(t_data *data, t_tuple origin, double radius);
 t_intersections		get_cylinder_intersections(t_object *cylinder, t_ray ray);
 t_tuple				cylinder_normal_at(t_object *cylinder, t_tuple object_point);
 t_intersections		get_cylinder_caps_intersections(t_object *cylinder, t_ray ray, t_intersections intersections);
 
 /* cone.c-------------------------------------------------------------------- */
 t_intersections	get_cone_intersections(t_object *cone, t_ray ray);
-t_object		*create_cone(t_data *data, t_tuple origin, int radius);
+t_object		*create_cone(t_data *data, t_tuple origin, double radius);
 t_tuple			cone_normal_at(t_object *cone, t_tuple object_point);
 
 /* material.c ----------------------------------------------------------------*/
